fetch and cache text of stories
useful for reading the summary of a page when the site is down, and
allows us to show the content of the story in the rss feed as well
as making its content searchable.

whether this is legal/responsible is up for debate, so this is not
being committed on master yet.